requirements.txt
python-telegram-bot
openai
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2